+++ Assets/HUF/Utils/Editor/BuildSupport/iOS/iOSProjectBaseFrameworkManager.cs
@@ -7,9 +7,23 @@ namespace HUF.Utils.BuildSupport.Editor.iOS
     public abstract class iOSProjectBaseFrameworkManager : iOSProjectBaseFixer
     {
         protected abstract IEnumerable<string> FrameworksToAdd { get; }
+#if UNITY_2019_3_OR_NEWER
+        protected virtual IEnumerable<string> FrameworksToAddToMainTarget { get; } = new string[0];
+#endif
 
-        protected override bool Process(PBXProject project, string targetGuid, string projectPath)
+        protected override bool Process(PBXProject project)
         {
+#if UNITY_2019_3_OR_NEWER
+            var mainTargetGuid = project.GetUnityMainTargetGuid();
+            foreach (var framework in FrameworksToAddToMainTarget)
+            {
+                TryAddFramework(project, mainTargetGuid, framework);
+            }
+            var targetGuid = project.GetUnityFrameworkTargetGuid();
+#else
+            string targetName = PBXProject.GetUnityTargetName();
+            var targetGuid = project.TargetGuidByName(targetName);
+#endif
             foreach (var framework in FrameworksToAdd)
             {
                 TryAddFramework(project, targetGuid, framework);