+++ Assets/HUF/Utils/Runtime/PauseManager.cs
@@ -29,10 +29,7 @@ namespace HUF.Utils.Runtime
         [PublicAPI]
         public event Action<ScreenOrientation> OnScreenOrientationChange;
 
-        /// <summary>
-        /// Raised when the game is paused.
-        /// </summary>
-        [PublicAPI]
+        [Obsolete( "Use `InteractionManager.Instance.OnBackButtonPress` instead." )]
         public event Action OnBackButtonPress;
 #pragma warning restore 0067
 
@@ -66,9 +63,14 @@ namespace HUF.Utils.Runtime
                 OnApplicationPause( false );
         }
 #endif
-        
+
         void Awake()
         {
+#if UNITY_ANDROID || UNITY_EDITOR || UNITY_STANDALONE || UNITY_STANDALONE_OSX
+#pragma warning disable 0618
+            InteractionManager.Instance.OnBackButtonPress += () => OnBackButtonPress.Dispatch();
+#pragma warning restore 0618
+#endif
             lastScreenOrientation = Screen.orientation;
 #if UNITY_EDITOR
             EditorApplication.pauseStateChanged += HandlePauseState;
@@ -82,10 +84,6 @@ namespace HUF.Utils.Runtime
                 lastScreenOrientation = Screen.orientation;
                 OnScreenOrientationChange.Dispatch( lastScreenOrientation );
             }
-#if UNITY_ANDROID
-            if (Input.GetKeyDown(KeyCode.Escape))
-                OnBackButtonPress.Dispatch();
-#endif
         }
     }
 }